+++ crates/vertigo-cli/src/serve/html/element.rs
@@ -151,11 +151,7 @@ impl AllElements {
 
     fn get_response_one_elements(&self, node_id: u64, with_id: bool) -> HtmlNode {
         let Some(node) = self.all.get(&node_id) else {
-            if node_id == 1 {
-                panic!("Can't find root <html> element");
-            } else {
-                unreachable!();
-            }
+            return HtmlNode::Comment("No <html> element".to_string())
         };
 
         match node {