+++ CHANGES.md
@@ -6,6 +6,7 @@
 ### Fixed
 
 * vertigo-cli: Don't html-escape styles embedded during SSR
+* vertigo-cli: Don't panic when missing root html element
 
 ## 0.4.2 - 2024-02-06